+++ R/posterior_traj.R
@@ -181,6 +181,7 @@
 #'   \code{\link{posterior_survfit}}
 #' 
 #' @examples
+#' if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
 #' \donttest{
 #'   # Run example model if not already loaded
 #'   if (!exists("example_jm")) example(example_jm)
@@ -270,7 +271,7 @@
 #'                         re.form = NA)
 #'   head(pt8)  # note the much narrower ci, compared with pt5
 #' }
-#' 
+#' }
 posterior_traj <- function(object, m = 1, newdata = NULL, newdataLong = NULL,
                            newdataEvent = NULL, interpolate = TRUE, extrapolate = FALSE,
                            control = list(), last_time = NULL, prob = 0.95, ids, 
@@ -283,6 +284,7 @@ posterior_traj <- function(object, m = 1, newdata = NULL, newdataLong = NULL,
   id_var    <- object$id_var
   time_var  <- object$time_var
   grp_stuff <- object$grp_stuff[[m]]
+  glmod     <- object$glmod[[m]]
   if (!is.null(seed)) 
     set.seed(seed)
   if (missing(ids)) 
@@ -397,7 +399,8 @@ posterior_traj <- function(object, m = 1, newdata = NULL, newdataLong = NULL,
     }
   }
     
-  ytilde <- posterior_predict(object, newdata = newX, m = m, stanmat = stanmat, ...)
+  newOffset <- model.offset(model.frame(terms(glmod), newX))
+  ytilde <- posterior_predict(object, newdata = newX, m = m, stanmat = stanmat, offset = newOffset, ...)
   if (return_matrix) {
     attr(ytilde, "mu") <- NULL # remove attribute mu
     return(ytilde) # return S * N matrix, instead of data frame 
@@ -493,7 +496,8 @@ posterior_traj <- function(object, m = 1, newdata = NULL, newdataLong = NULL,
 #' @seealso \code{\link{posterior_traj}}, \code{\link{plot_stack_jm}},
 #'   \code{\link{posterior_survfit}}, \code{\link{plot.survfit.stanjm}}   
 #'     
-#' @examples 
+#' @examples
+#' if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
 #' \donttest{
 #'   # Run example model if not already loaded
 #'   if (!exists("example_jm")) example(example_jm)
@@ -525,6 +529,7 @@ posterior_traj <- function(object, m = 1, newdata = NULL, newdataLong = NULL,
 #'   plot1 + 
 #'     ggplot2::theme(strip.background = ggplot2::element_blank()) +
 #'     ggplot2::labs(title = "Some plotted longitudinal trajectories")
+#' }
 #' }
 plot.predict.stanjm <- function(x, ids = NULL, limits = c("ci", "pi", "none"), 
                                 xlab = NULL, ylab = NULL, vline = FALSE,