+++ tests/testthat.R
@@ -19,6 +19,8 @@ library(testthat)
 library(rstanarm)
 Sys.unsetenv("R_TESTS")
 example(example_model)
-test_check("rstanarm", invert = FALSE,
-           filter = if (Sys.getenv("NOT_CRAN") != "true") "stan_functions")
+if (.Platform$OS.type != "windows" || .Platform$r_arch != "i386") {
+  test_check("rstanarm", invert = FALSE,
+             filter = if (Sys.getenv("NOT_CRAN") != "true") "stan_functions")
+}